Xcode 12 Workaround

+++ Sources/LsbBitWriter.swift
@@ -80,7 +80,7 @@ public final class LsbBitWriter: BitWriter {
                    "\(signedNumber) is too big to be encoded by biased representation with \(bias) bias using \(bitsCount) bits")
             self.write(unsignedNumber: encoded, bitsCount: bitsCount)
         case .radixNegativeTwo:
-            let mask: UInt
+            let mask: UInt64
             if MemoryLayout<UInt>.size == 8 {
                 mask = 0xAA_AA_AA_AA_AA_AA_AA_AA
             } else if MemoryLayout<UInt>.size == 4 {
@@ -88,12 +88,12 @@ public final class LsbBitWriter: BitWriter {
             } else {
                 fatalError("Unknown Int bit width")
             }
-            let unsignedBitPattern = UInt(bitPattern: signedNumber)
+            let unsignedBitPattern = UInt64(bitPattern: Int64(signedNumber))
             let encoded = (unsignedBitPattern &+ mask) ^ mask
             let encodedBound = bitsCount == UInt.bitWidth ? UInt.max : ((1 << bitsCount) - 1)
             assert(encoded <= encodedBound,
                    "\(signedNumber) will be truncated when represented by -2 radix using \(bitsCount) bits")
-            self.write(unsignedNumber: encoded, bitsCount: bitsCount)
+            self.write(unsignedNumber: UInt(encoded), bitsCount: bitsCount)
         default:
             fatalError("Not implemented")
         }
